backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.0
redis[hiredis]==5.1.0
pydantic==2.9.0
email-validator>=2.0
pydantic-settings==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.0
cryptography==43.0.0
pqcrypto==0.1.3
python-multipart==0.0.12
celery[redis]==5.4.0
structlog==24.4.0
prometheus-client==0.21.0
tiktoken==0.7.0
numpy>=1.26
scikit-learn>=1.5
